# improve std.zig.parse performance using flat arrays for AST nodes and tokens