std.ast uses a singly linked list for lists of things. This is a
breaking change to the self-hosted parser API.

std.ast.Tree has been separated into a private "Parser" type which
represents in-progress parsing, and std.ast.Tree which has only
"output" data. This means cleaner, but breaking, API for parse results.
Specifically, `tokens` and `errors` are no longer SegmentedList but a
slice.

The way to iterate over AST nodes has necessarily changed since lists of
nodes are now singly linked lists rather than SegmentedList.

From these changes, I observe the following on the
self-hosted-parser benchmark from ziglang/gotta-go-fast:

with std.heap.page_allocator as the general purpose allocator:
throughput: 45.6 MiB/s => 55.6 MiB/s
maxrss: 359 KB => 342 KB

with std.heap.c_allocator as the general purpose allocator:
throughput: 46.2 MiB/s => 76.4 MiB/s
maxrss: 355 KB => 392 KB

This commit breaks the build; more updates are necessary to fix API
usage of the self-hosted parser.